ssr-studio/pyproject.toml
[project]
name = "ssr-studio"
version = "1.0.0"
description = "Self-Play SWE-RL Demo & Research Platform"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "SSR Studio Team"}
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "docker>=7.0.0",
    "httpx>=0.26.0",
    "python-multipart>=0.0.6",
    "boto3>=1.34.0",
    "aiofiles>=23.2.0",
    "redis>=5.0.0",
    "celery>=5.3.0",
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    "tiktoken>=0.5.0",
    "unidiff>=0.7.5",
    "gitpython>=3.1.40",
    "pyyaml>=6.0.1",
    "structlog>=24.1.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
training = [
    "torch>=2.1.0",
    "transformers>=4.37.0",
    "peft>=0.8.0",
    "trl>=0.7.0",
    "accelerate>=0.26.0",
    "bitsandbytes>=0.42.0",
    "vllm>=0.3.0",
]

[project.scripts]
ssr = "ssr_studio.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ssr_studio"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
